requirements.txt
numpy>=2.0.0
torch==2.5.1
torchvision>=0.2.0
pillow==11.0.0
